% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.R
\name{forecast_FRED}
\alias{forecast_FRED}
\title{Generate forecast of FRED data}
\usage{
forecast_FRED(df_in, fred_id, freq, h)
}
\arguments{
\item{df_in}{A data frame with FRED data}

\item{fred_id}{yep}

\item{freq}{yep}

\item{h}{yep}
}
\value{
A data.frame
}
\description{
Title duh
}
