% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_FRED}
\alias{get_FRED}
\title{Get FRED Data into DF}
\usage{
get_FRED(fred_id, years_history)
}
\arguments{
\item{fred_id}{ID for FRED data ticker}

\item{years_history}{NUmber of hostoirial years of data to grab who car about splenbign}
}
\description{
See title
}
