% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/out.R
\name{write_sqlite}
\alias{write_sqlite}
\title{Write out forecast data}
\usage{
write_sqlite(df_fcast, fred_id)
}
\arguments{
\item{df_fcast}{The data frame of forecasted FRED data as returned by \code{\link[=forecast_FRED]{forecast_FRED()}}}

\item{fred_id}{yep}
}
\value{
Nothing
}
\description{
Write out forecast data
}
