% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/in.R
\name{get_FRED}
\alias{get_FRED}
\title{Get FRED Data}
\usage{
get_FRED(fred_id)
}
\arguments{
\item{fred_id}{FRED series ID to fetch}
}
\value{
A data frame
}
\description{
This function returns FRED data as a data frame for further analysis
}
